+++ docs/conf.py
@@ -94,6 +94,7 @@ extlinks = {
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3', None),
     'requests': ('https://requests.readthedocs.io/en/master/', None),
+    'requests_cache': ('https://requests-cache.readthedocs.io/en/stable/', None),
     'urllib3': ('https://urllib3.readthedocs.io/en/stable/', None),
 }